backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain-core==0.3.69
langchain==0.3.26
langchain-openai==0.3.28
langgraph==0.5.3
langchain-community==0.3.27
chromadb==0.4.24
sse-starlette==1.6.1
aiohttp==3.10.11
pydantic-settings==2.10.1
starlette==0.40.0

# Persistent checkpoint backend
langgraph-checkpoint-sqlite>=1.0.3
